refactor(lifted-air,lifted-stark): address PR review comments
- move the MultiAir trait + ReductionError into air.rs; rename
  instance.rs → statement.rs (now centered on Statement/ProverStatement)
- spell out concrete `*Ma` test/bench types as `*MultiAir`
- drop the `BUG:` prefix from assert/expect messages in favor of
  idiomatic phrasing
- add ConstraintDegrees::max(); keep log_quotient_degree a free function
  using it
- simplify the README protocol/ordering blocks; clarify the
  num_aux_values, eval_external, and verifier docs
