+++ stark/miden-lifted-air/src/air.rs
@@ -1,4 +1,5 @@
-//! The `LiftedAir` super-trait for AIR definitions in the lifted STARK system.
+//! AIR traits for the lifted STARK system: [`LiftedAir`] (a single AIR) and
+//! [`MultiAir`] (the AIR collection plus the cross-AIR behavior on it).
 //!
 //! # Panic safety of `eval()`
 //!
@@ -20,10 +21,11 @@
 //! Use [`crate::debug::check_builder_shape`] to verify a concrete builder's
 //! dimensions match the AIR before calling `eval()`.
 
-use alloc::vec::Vec;
+use alloc::{boxed::Box, vec::Vec};
 
 use p3_air::BaseAir;
-use p3_field::Field;
+use p3_challenger::CanObserve;
+use p3_field::{ExtensionField, Field};
 use p3_matrix::dense::RowMajorMatrix;
 
 use crate::{
@@ -86,8 +88,11 @@ pub trait LiftedAir<F: Field, EF>: Sync + BaseAir<F> {
     ///
     /// These are the values returned by
     /// [`MultiAir::build_aux_traces`](crate::MultiAir::build_aux_traces)
-    /// alongside the aux trace matrix. Their count may differ from
-    /// [`aux_width`](Self::aux_width) (the number of aux trace columns).
+    /// alongside the aux trace matrix. Their count is independent of
+    /// [`aux_width`](Self::aux_width) (the number of aux trace columns): aux
+    /// columns carry per-row witness data, whereas aux values are a fixed set
+    /// of scalars (e.g. a logup running-sum total) bound once into the
+    /// transcript.
     ///
     /// These values are exposed to AIR constraints as *permutation values* via
     /// [`PermutationAirBuilder::permutation_values`](crate::PermutationAirBuilder::permutation_values).
@@ -138,7 +143,7 @@ pub trait LiftedAir<F: Field, EF>: Sync + BaseAir<F> {
 ///
 /// `base` is the maximum degree multiple over the base-field constraints and
 /// `ext` over the extension-field constraints (each `0` if the AIR has none of
-/// that kind). Consumers that need a single value take `base.max(ext)`; the
+/// that kind). Consumers that need a single value take [`max`](Self::max); the
 /// split is exposed via [`LiftedAir::constraint_degree`] so a per-AIR override
 /// can be sharp without redoing the symbolic pass.
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
@@ -147,6 +152,14 @@ pub struct ConstraintDegrees {
     pub base: usize,
     /// Max degree multiple over extension-field constraints (`0` if there are none).
     pub ext: usize,
+}
+
+impl ConstraintDegrees {
+    /// The combined degree multiple: the larger of [`base`](Self::base) and
+    /// [`ext`](Self::ext).
+    pub fn max(&self) -> usize {
+        self.base.max(self.ext)
+    }
 }
 
 /// Raw constraint degree multiples from symbolic evaluation, before any
@@ -175,4 +188,131 @@ where
         .max()
         .unwrap_or(0);
     ConstraintDegrees { base, ext }
+}
+
+/// Boxed error returned by [`MultiAir::eval_external`].
+pub type ReductionError = Box<dyn core::error::Error + Send + Sync>;
+
+// ============================================================================
+// MultiAir trait
+// ============================================================================
+
+/// The circuit for a multi-AIR statement: the AIR collection plus the
+/// cross-AIR behavior that operates on it.
+///
+/// Methods take `&self` so an impl can carry the AIRs and any protocol-level
+/// state (closures, lookup tables, shared parameters). Because the AIRs are
+/// owned here, a `MultiAir` is constructed per proof.
+///
+/// The framework defines instance order as the position of an AIR within
+/// [`Self::airs`]. Every per-AIR slice elsewhere on [`Statement`](crate::Statement) /
+/// [`ProverStatement`](crate::ProverStatement) uses the same ordering.
+pub trait MultiAir<F, EF>
+where
+    F: Field,
+    EF: ExtensionField<F>,
+{
+    /// AIR type. Heterogeneous AIRs are expressed via caller-defined enum
+    /// wrappers, exactly as before.
+    type Air: LiftedAir<F, EF>;
+
+    /// The AIRs in instance order — the single source of that ordering.
+    fn airs(&self) -> &[Self::Air];
+
+    /// Upper bound on `aux_inputs().len()` accepted by [`Self::eval_external`].
+    ///
+    /// Validated by [`Statement::new`](crate::Statement::new) before any
+    /// cryptographic work. Default `0`; implementations that consume
+    /// `aux_inputs` must override so the budget matches the schema their
+    /// `eval_external` decodes.
+    fn max_aux_inputs(&self) -> usize {
+        0
+    }
+
+    /// Cross-AIR external assertions.
+    ///
+    /// Returns a flat vector of extension-field values, each of which must
+    /// equal zero for the proof to be accepted. The caller (verifier) checks
+    /// each value against zero — individually, or batched via a random linear
+    /// combination — and is responsible for ensuring the supplied inputs match
+    /// what the AIRs describe.
+    ///
+    /// # Arguments
+    /// - `challenges`: shared extension-field challenge pool; each AIR consumes the prefix of
+    ///   length `air.num_randomness()`.
+    /// - `air_inputs`, `aux_inputs`: the inputs from the [`Statement`](crate::Statement).
+    /// - `aux_values`: per-AIR aux values in instance order. `aux_values[i]` belongs to
+    ///   `self.airs()[i]`.
+    /// - `log_trace_heights`: per-AIR log₂ trace heights in instance order.
+    ///
+    /// Default: refuses to be called with non-empty `aux_inputs`; otherwise
+    /// emits no assertions.
+    fn eval_external(
+        &self,
+        challenges: &[EF],
+        air_inputs: &[F],
+        aux_inputs: &[F],
+        aux_values: &[&[EF]],
+        log_trace_heights: &[u8],
+    ) -> Result<Vec<EF>, ReductionError> {
+        if !aux_inputs.is_empty() {
+            return Err("default `eval_external` received non-empty `aux_inputs` — override \
+                 `eval_external` to consume them"
+                .into());
+        }
+        let _ = (challenges, air_inputs, aux_values, log_trace_heights);
+        Ok(Vec::new())
+    }
+
+    /// Build every AIR's auxiliary trace and aux values.
+    ///
+    /// # Arguments
+    /// - `traces`, `air_inputs`, `aux_inputs`: instance-order slices owned by the
+    ///   [`ProverStatement`](crate::ProverStatement) / [`Statement`](crate::Statement).
+    /// - `challenges`: sized to the maximum `num_randomness()` across AIRs; each AIR consumes the
+    ///   prefix matching its own `num_randomness()`.
+    ///
+    /// # Returns
+    /// `(aux_traces, aux_values)` in instance order — one entry per AIR.
+    /// `aux_traces[i]` has width `self.airs()[i].aux_width()` and height matching
+    /// `traces[i]`; `aux_values[i]` has length `self.airs()[i].num_aux_values()`.
+    fn build_aux_traces(
+        &self,
+        traces: &[&RowMajorMatrix<F>],
+        air_inputs: &[F],
+        aux_inputs: &[F],
+        challenges: &[EF],
+    ) -> (Vec<RowMajorMatrix<EF>>, Vec<Vec<EF>>);
+
+    /// Absorb per-proof state into the Fiat-Shamir challenger.
+    ///
+    /// Default order: `air_inputs`, then `aux_inputs`, then each log trace
+    /// height in instance order. Overrides must preserve this ordering unless
+    /// they account for the change on both prover and verifier.
+    ///
+    /// # Soundness gap (TODO)
+    ///
+    /// The default binds inputs and trace heights but does NOT canonically
+    /// bind the `MultiAir` itself — neither its AIR collection nor
+    /// `eval_external` — into Fiat-Shamir. Until the symbolic-graph binding
+    /// lands (tracked in <https://github.com/0xMiden/crypto/issues/970>),
+    /// callers MUST observe the `MultiAir`'s AIR configurations into the
+    /// challenger before calling the prover or verifier.
+    fn observe<C: CanObserve<F>>(
+        &self,
+        challenger: &mut C,
+        air_inputs: &[F],
+        aux_inputs: &[F],
+        log_trace_heights: &[u8],
+    ) {
+        for &v in air_inputs {
+            challenger.observe(v);
+        }
+        for &v in aux_inputs {
+            challenger.observe(v);
+        }
+        for &h in log_trace_heights {
+            challenger.observe(F::from_u8(h));
+        }
+    }
 }
