+++ stark/miden-lifted-stark/benches/per_air_degree_opt.rs
@@ -182,9 +182,9 @@ impl<A: LiftedAir<Felt, QuadFelt>> LiftedAir<Felt, QuadFelt> for OverrideConstra
 // MultiAir (trivial: constant-challenge aux column for each trace)
 // -----------------------------------------------------------------------------
 
-struct BenchMa;
+struct BenchMultiAir;
 
-impl MultiAir<Felt, QuadFelt> for BenchMa {
+impl MultiAir<Felt, QuadFelt> for BenchMultiAir {
     type Air = OverrideConstraintDegree<BenchAir>;
 
     fn build_aux_traces(
@@ -272,7 +272,7 @@ fn run_prove(
     let chip_trace = generate_trace(BenchAirKind::Chip.recurrence_power(), chip_height);
 
     let statement =
-        Statement::new(BenchMa, vec![core_air, chip_air], Vec::new(), Vec::new()).unwrap();
+        Statement::new(BenchMultiAir, vec![core_air, chip_air], Vec::new(), Vec::new()).unwrap();
     let prover_statement = ProverStatement::new(statement, vec![core_trace, chip_trace]).unwrap();
 
     eprintln!("\n{}", "=".repeat(70));
