+++ stark/miden-lifted-stark/src/testing/test_per_air_degree.rs
@@ -132,17 +132,17 @@ impl LiftedAir<Felt, QuadFelt> for PeriodicPowerAir {
 // MultiAir: trivial constant-challenge aux column for each trace.
 // ---------------------------------------------------------------------------
 
-struct TwoTraceMa<A> {
+struct TwoTraceMultiAir<A> {
     airs: Vec<A>,
 }
 
-impl<A> TwoTraceMa<A> {
+impl<A> TwoTraceMultiAir<A> {
     fn new(airs: Vec<A>) -> Self {
         Self { airs }
     }
 }
 
-impl<A> MultiAir<Felt, QuadFelt> for TwoTraceMa<A>
+impl<A> MultiAir<Felt, QuadFelt> for TwoTraceMultiAir<A>
 where
     A: LiftedAir<Felt, QuadFelt>,
 {
@@ -197,7 +197,7 @@ fn run_upsample_case(low_power: u64, low_height: usize, high_power: u64, high_he
     let t_high = generate_pow_trace(high_power, Felt::from_u64(11), high_height);
 
     let statement =
-        Statement::new(TwoTraceMa::new(vec![low, high]), Vec::new(), Vec::new()).unwrap();
+        Statement::new(TwoTraceMultiAir::new(vec![low, high]), Vec::new(), Vec::new()).unwrap();
     let prover_statement = ProverStatement::new(statement, vec![t_low, t_high]).unwrap();
     prove_and_verify_statement(&prover_statement);
 }
@@ -226,7 +226,7 @@ fn upsample_fires_with_periodic_columns() {
     let t_high = generate_pow_trace(5, Felt::from_u64(11), 16);
 
     let statement =
-        Statement::new(TwoTraceMa::new(vec![low, high]), Vec::new(), Vec::new()).unwrap();
+        Statement::new(TwoTraceMultiAir::new(vec![low, high]), Vec::new(), Vec::new()).unwrap();
     let prover_statement = ProverStatement::new(statement, vec![t_low, t_high]).unwrap();
     prove_and_verify_statement(&prover_statement);
 }
